Prevent exceptions in get_object when objects does not exists (keeps logfile clean)

+++ lib/fakes3/file_store.rb
@@ -96,6 +96,7 @@ module FakeS3
         get_or_create_bucket(bucket)
         real_obj = S3Object.new
         obj_root = File.join(@root,bucket,object_name,FAKE_S3_METADATA_DIR)
+        return nil unless File.exist? obj_root
         metadata = File.open(File.join(obj_root, "metadata")) { |file| YAML::load(file) }
         real_obj.name = object_name
         real_obj.md5 = metadata[:md5]
